+++ src/BecklynRadBundle.php
@@ -2,10 +2,8 @@
 
 namespace Becklyn\Rad;
 
+use Becklyn\Rad\DependencyInjection\BecklynRadExtension;
 use Becklyn\Rad\DependencyInjection\DoctrineExtensionsCompilerPass;
-use Becklyn\Rad\Doctrine\Types\SerializedType;
-use Becklyn\RadBundles\Bundle\BundleExtension;
-use Doctrine\DBAL\Types\Type;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -19,7 +17,7 @@ class BecklynRadBundle extends Bundle
      */
     public function getContainerExtension ()
     {
-        return new BundleExtension($this);
+        return new BecklynRadExtension($this);
     }
 
 
@@ -37,16 +35,5 @@ class BecklynRadBundle extends Bundle
     public function getPath () : string
     {
         return \dirname(__DIR__);
-    }
-
-
-    /**
-     * @inheritDoc
-     */
-    public function boot () : void
-    {
-        parent::boot();
-
-        Type::addType(SerializedType::NAME, SerializedType::class);
     }
 }